generator/requirements.txt
google-genai==0.2.2
python-dotenv==1.0.0
